Comment UI and API

+++ app/src/main/java/com/tripleThreads/taxiyaz/repository/RouteRepository.kt
@@ -1,9 +1,11 @@
 package com.tripleThreads.taxiyaz.repository
 
+import android.util.Log
 import androidx.annotation.WorkerThread
 import androidx.lifecycle.LiveData
 import com.tripleThreads.taxiyaz.data.route.RouteDao
 import com.tripleThreads.taxiyaz.data.route.Route
+import com.tripleThreads.taxiyaz.data.APIHelpers.RouteAPI
 import com.tripleThreads.taxiyaz.network.RouteService
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
@@ -18,21 +20,42 @@ class RouteRepository(private val dao: RouteDao, private val routeService: Route
 
     @WorkerThread
     fun insert(route: Route) {
+        if(insertRouteToAPI(route)) {
+            dao.insertRoute(route)
+        }
+        else{
+            //connection error
+        }
+    }
+    @WorkerThread
+    fun cache(route: Route) {
         dao.insertRoute(route)
+
     }
 
     @WorkerThread
     fun update(route: Route) {
-        dao.updateRoute(route)
+        if(updateRouteInAPI(route)) {
+            dao.updateRoute(route)
+        }
+        else{
+            //connection error
+        }
     }
 
     @WorkerThread
     fun delete(route: Route) {
-        dao.deleteRoute(route)
+        if(deleteRoutefromAPI(route)) {
+            dao.deleteRoute(route)
+        }
+        else{
+            //connection error
+        }
     }
 
     @WorkerThread
-    fun getByName(name: String): LiveData<Route> {
+    fun getByName(name: String): LiveData<List<Route>> {
+        Log.d("check", "In repo1")
         getRouteFromAPI(name)
         allRoutes = dao.getAllRoutes()
         return dao.getRouteByName(name)
@@ -46,24 +69,62 @@ class RouteRepository(private val dao: RouteDao, private val routeService: Route
     //network functions
 
     @WorkerThread
-    fun getRouteFromAPI(title: String) {
-        GlobalScope.launch(Dispatchers.IO) {
+    fun getRouteFromAPI(title: String){
+            GlobalScope.launch(Dispatchers.IO) {
 
-            if (routeService != null) {
-                val response: Response<List<Route>> = routeService.getAllRoutesAsync().await()
-                val routes = response.body()
+                Log.d("check", "In repo2")
+                    if (routeService != null) {
+                        val response: Response<List<RouteAPI>> = routeService.getRouteByTitleAsync(title)!!.await()
+                        val routes = response.body()
+                        Log.d("check", "In repo3")
 
-                if (routes != null) {
-                    withContext(Dispatchers.IO) {
-                        routes.forEach { route -> insert(route) }
+                        if (routes != null) {
+                            withContext(Dispatchers.IO) {
+                                dao.deleteAll()
+                                routes.forEach { route -> cache(route.convertToRoute())}
+                            }
+                        }
                     }
-                }
-            } else {
+                    else{
+
+                    }
 
             }
+    }
 
+    @WorkerThread
+    fun insertRouteToAPI(route: Route): Boolean {
+        var added= false
+        GlobalScope.launch(Dispatchers.IO) {
+            if(routeService != null){
+                routeService.addRoute(route)
+                added = true
+            }
         }
+        return added
+    }
 
+    @WorkerThread
+    fun updateRouteInAPI(route: Route): Boolean {
+        var edited= false
+        GlobalScope.launch(Dispatchers.IO) {
+            if(routeService != null){
+                routeService.editRoute(route.routeId.toLong(),route)
+                edited = true
+            }
+        }
+        return edited
+    }
 
+    @WorkerThread
+    fun deleteRoutefromAPI(route: Route): Boolean {
+        var deleted= false
+        GlobalScope.launch(Dispatchers.IO) {
+            if(routeService != null){
+                routeService.deleteRoute(route.routeId.toLong())
+                deleted = true
+            }
+        }
+        return deleted
     }
 }
